chore: align structs and use of pointers to reduce space usage
Some of the structs were not aligned and therefore using some more bytes
and also using more cachelines that were actually needed.

Make Pod variable inside PostgresqlStatus a pointer to reduce the size
of the struct in 1056 bytes reducing the amount of space that we move
around everytime we query the status of the PostgreSQL cluster.

Closes #2457

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>